reset middleware pipeline counter on exit
When making an API call that goes down the
http/go/nethttp/pipeline.go Next()->middleware.Intercept()->Next() loop,
the pipeline.middlewareIndex variable keeps track to make sure the next
Intercept() call is made in the chain.

After all of them have done their work and the results are being
returned to the original API caller, the middlewareIndex is left at
non-zero, so subsequent calls down the pipeline.go path will fail to
make all the middleware.Intercept() calls as middlewareIndex is not less
than len(middlewares).

Add a defer so that the index gets reset to zero when exiting the
function call chain so that sequential passes down the pipeline chain
can behave like the very first call.