Convert Java package names to lower case after escaping
When a reserved keyword is used in a place where it will put in a
package name, the keyword is escaped like: new -> newEscaped. However,
the package names are converted to lower case prior to this conversion,
resulting in package names with an upper case character. This is not
idiomatic Java and also causes compilation issues, because the package
directory is created in lower case. The conversion to lower case must be
done after escaping keywords.