Do not convert first char to upper on type names during code model construction
Types are sometimes created with an upper case type name and sometimes
with a lower case type name. This has no functional effect on the
resulting code, because the writers will do this, but the converted name
can end up in the comment. The comment should follow the spec and should
be stable as much as possible.