Copied Dart refiner code over original refiner
Previously we used a dart refiner we created from scratch.
Just to be sure it only contains the code we actually need.
Currently the differences between the existing refiner and the newly created refiner are small.
At this point we decided to copy our refiner code over the existing refiner.